Reverting minions and doom-modeline
doom-modeline has some visual artifacts for me when using helm and
sometimes when opening a help buffer.

Minions doesn't look so nice with smart-modeline. It ends up taking
over the modes display bit completely and I loose the coloring for the
major mode name. I don't want to bother tweaking it right now and I'm
back with delight since it was already tweaked in a way that was ok
for me and it's easy to add to new modes with use-package.